requirements.txt
# PDF 处理
PyMuPDF==1.23.8

# 文本处理和嵌入
sentence-transformers==2.3.1

# 机器学习
xgboost==2.0.3
scikit-learn==1.4.0
imbalanced-learn==0.12.0

# 深度学习（仅用于图像特征提取）
torch==2.1.2
torchvision==0.16.2

# 数据处理
pandas==2.1.4
numpy==1.26.3

# 图像处理
Pillow==10.2.0

# 网页爬取
requests==2.31.0
beautifulsoup4==4.12.3

# 配置文件
PyYAML==6.0.1

# 可视化（可选）
matplotlib==3.8.2
seaborn==0.13.1

# 进度条
tqdm==4.66.1
